--- cache/30731291f0b752ae4ba00b55b2fd560a.json
@@ -1,1 +1,0 @@
-{"hour":"2014-02-25 10:00:00","min_price":"0.00085000","max_price":"0.00120000","avg_price":"0.00105","units_traded":"17144945.37121158","number_of_trades":"186","opening_price":"0.00102000","closing_price":"0.00111000","date":1393412301,"volume_1day":"17144945.37121158","volume_7day":"82052589.10250218","CacheHit":1}
