--- cache/45a82579528dcdc6535fdcb693a36170.json
@@ -1,89 +1,0 @@
-["http://destiny.gg",
-"http://reddit.com/r/destiny",
-"www.google.com",
-"meatspin.com",
-"bing.com",
-"shoveldog",
-"https://www.youtube.com/watch?v=cq7UtwkvNMc&feature=youtu.be",
-"antidisestablishmentternianism",
-"http://i.imgur.com/RmbI0GN.gif",
-"http://i.imgur.com/9gOsF8r.gif",
-"http://www.reddit.com/r/soapyanimals/",
-"http://i.imgur.com/MrSXmd5.gif",
-"http://i.imgur.com/dvd88mv.gif?1",
-"http://tehurn.com/frankerz",
-"sexual",
-"http://i.imgur.com/ftRaIca.gif",
-"Well",
-"༼",
-"http://i.imgur.com/mVcpG7O.png",
-"I",
-"http://www.eveonline.com/",
-"Destiny",
-"#BoneMom",
-"http://i.imgur.com/X0khkSD.png",
-"So",
-"https://www.youtube.com/watch?v=DUgoUGncBmY",
-"Already",
-"There",
-"https://www.youtube.com/watch?v=oHg5SJYRHA0",
-"';shutdown",
-"DuckerZ",
-"Honestly,",
-"!submit",
-"I",
-"are.you.rustled.RTBA.gg",
-"http://www.staggeringbeauty.com/",
-"http://i.imgur.com/8qXX7qk.jpg",
-"http://www.youtube.com/watch?v=5hfYJsQAhl0",
-"http://www.chickenonaraft.com/",
-"http://www.pornmd.com/live-search",
-"http://www.omfgdogs.com/",
-"http://www.pornmd.com/live-search",
-"http://(__)_)============D.com",
-"https://www.youtube.com/watch?v=7yBnl_krN_U",
-"this",
-"http://www.dartmouth.edu/~milton/reading_room/pl/book_1/",
-"http://zombo.com/",
-"https://www.youtube.com/watch?v=bv7vPtXdWnk",
-"http://www.lingscars.com/",
-"https://www.youtube.com/watch?v=0bDY4vrcelw",
-"faggotry",
-"https://www.youtube.com/watch?v=oHg5SJYRHA0",
-"https://www.youtube.com/watch?v=3_vmOvfbg90",
-"https://www.youtube.com/watch?v=oHg5SJYRHA0",
-"http://grooveshark.com/s/Ember+Flareon+Theme+Of+Lavender+Town/5msYlF?src=5",
-"http://i.imgur.com/kiakPwg.gif",
-"https://www.youtube.com/watch?v=DUgoUGncBmY",
-"http://i.imgur.com/CJMHk1s.gif",
-"http://www.youtube.com/watch?v=Cgoqrgc_0cM",
-"http://i.minus.com/iKxmjHgQHXjOC.gif",
-"http://i.imgur.com/lZrcZ2P.gif",
-"to",
-"http://i.imgur.com/8SLSWsg.jpg",
-"http://i.imgur.com/sDHgeuI.jpg",
-"http://boards.4chan.org/mlp/",
-"http://imgur.com/r/195/XJKSGfo",
-"You'reAFaggot",
-"psuedointellectualfaggot.com",
-"YourMom'sTits",
-"YourDad'sDick",
-"NoTears",
-"http://i.imgur.com/PE88Xr5.jpg",
-"http://www.youtube.com/watch?v=D6AzUv3LXDc",
-"http://grooveshark.com/s/By+Myslf/594Y81?src=5",
-"What.the.fuck.did.you.just.fucking.say.about.me.you.little.bitch.Ill.have.you.know.I.graduated.top.of.my.class.in.the.Navy.Seals",
-"it.belongs.in.a.museum",
-"D0GE",
-"!doge",
-"https://www.youtube.com/watch?v=lH_XwLZ6Y0M",
-"What.the.fuck.did.you.just.fucking.say.about.me.you.little.bitch.Ill.have.you.know.I.graduated.top.of.my.class.in.the.Navy.Seals.gg",
-"http://grooveshark.com/s/When+They+Come+For+Me/39zR3b?src=5",
-"https://www.youtube.com/watch?v=lH_XwLZ6Y0M",
-"OverRustle",
-"http://imgur.com/gpqEtzV",
-"https://www.youtube.com/watch?v=FmWOm1rNFHQ",
-"https://www.youtube.com/watch?v=mSBKZPNsK74",
-"http://imgur.com/c0BeE",
-"http://flapmmo.com/",
-"maybe.you.would.have.held.your.fucking.tongue.But.you.couldn.t.you.didn.t.and.now.you.re.paying.the.price.you.goddamn.idiot.I.will.shit.fury.all.over.you.and.you.will.drown.in.it.You.re.fucking.dead.kiddo.gg"]
