+++ safe_cache.rb
@@ -1,0 +1,33 @@
+module SafeCache
+  CACHE_DURATION = 60 #seconds
+  APP_ROOT = File.expand_path(File.dirname(__FILE__))
+  CACHE_FILE = APP_ROOT+"/cache/"
+  # safe cache! won't die if the bot dies
+  def is_expired?(cache)
+    cache["date"].to_i < (Time.now.to_i - CACHE_DURATION)
+  end
+  
+  def getcached(url)
+    _cached = instance_variable_get "@cached_#{hashed(url)}"
+    return _cached unless _cached.nil?
+    path = CACHE_FILE + "#{hashed(url)}.json"
+    if File.exists?(path)
+      f = File.open(path)
+      _cached = JSON.parse(f.read)
+      instance_variable_set("@cached_#{hashed(url)}", _cached)
+      return _cached
+    end
+    return nil
+  end
+  def setcached(url, jsn)
+    instance_variable_set("@cached_#{hashed(url)}", jsn)
+    path = CACHE_FILE + "#{hashed(url)}.json"
+    File.open(path, 'w') do |f2|
+      f2.puts JSON.unparse(jsn)
+    end
+  end
+
+  def hashed(url)
+    return Digest::MD5.hexdigest(url).to_s
+  end
+end
